// src/enums.ts
export enum Extension {
    KEYS = '.json',
    VALUES = '.v',
}

export enum File {
    STORE = 'store.json'
}

export enum Folder {
    KEYS = 'keys',
    VALUES = 'values',
}

export enum Value {
    TRUE = 1,
    FALSE = 0,
}

export enum Field {
    ID = 'i',
    TYPE = 't',
    DELETED = 'd',
    CREATED_AT = 'c',
    UPDATED_AT = 'u',
}